requirements.txt
streamlit==1.25.0
Pillow==9.3.0
pandas==2.1.1
requests==2.31.0
spacy==3.6.0

